Calculates a range of statistics of pedigrees

Statistics are those that will hopefully be useful for describing pedigrees to be used in quantitative genetic analyses of natural populations. This module will be most useful when cohort affinities for all individuals can be provided. All outputs are produced in a numerical form as well as in graphical summaries.

Usage

ped_stats(
  Ped,
  cohorts = NULL,
  dat = NULL,
  retain = "informative",
  includeA = TRUE,
  lowMem = FALSE
)

Arguments

Ped

A pedigree

cohorts

(Optional) Cohort affinities for members of the pedigree

dat

(Optional) Available data based upon which the pedigree can be pruned for just informative individuals

retain

The default value ('informative') results in pedigree being pruned to only those individuals who's records contribute to estimation of quantitative genetic parameters with respect to the available data specified in dat. Otherwise, specifying a value of 'ancestors' will result in the inclusion of all ancestors of phenotyped individuals.

includeA

If TRUE, additive genetic relatedness matrix is returned.

lowMem

If TRUE, then stats based on calculation of A are not performed.

Value

totalMaternities

Total number of maternities defined by the pedigree.

totalPaternities

Total number of paternities defined by the pedigree.

totalFullSibs

Total number of pair-wise full sib relationships defined by the pedigree.

totalMaternalSibs

Total number of pair-wise maternal sib relationships defined by the pedigree. To get the number of maternal half sibs, subtract totalFullSibs.

totalPaternalSibs

Total number of pair-wise paternal sib relationships defined by the pedigree. To get the number of paternal half sibs, subtract totalFullSibs.

totalMaternalGrandmothers

Total number of maternal grandmothers defined by the pedigree.

totalMaternalGrandfathers

Total number of maternal grandfathers defined by the pedigree.

totalPaternalGrandmothers

Total number of paternal grandmothers defined by the pedigree.

totalPaternalGrandfathers

Total number of paternal grandfathers defined by the pedigree.

pedigreeDepth

The pedigree depth, i.e. maximum number of ancestral generations, for each individual.

inbreedingCoefficients

Individual inbreeding coefficients

maternalSibships

Sibship size of each individual appearing the the dam column of the pedigree.

paternalSibships

Sibship size of each individual appearing the the sire column of the pedigree.

cumulativeRelatedness

Proportion of pair-wise relatedness values less than values ranging from 0 to 1.

relatednessCategories

Discretized distribution of relatedness.

analyzedPedigree

Returns the pedigree.

sampleSizesByCohort

(Optional) Number of individuals belonging to each cohort.

maternitiesByCohort

(Optional) Number of assigned maternities by offspring cohort.

paternitiesByCohort

(Optional) Number of assigned paternities by offspring cohort.

fullSibsByCohort

(Optional) Number of pair-wise full sib relationships by cohort - note the sum of these need not be equal to totalFullSibs in pedigrees of long-lived organisms.

maternalSibsByCohort

(Optional) Number of pair-wise maternal sib relationships by cohort - note the sum of these need not be equal to totalMaternalSibs in pedigrees of long-lived organisms.

paternalSibsByCohort

(Optional) Number of pair-wise paternal sib relationships by cohort - note the sum of these need not be equal to totalPaternalSibs in pedigrees of long-lived organisms.

maternalGrandmothersByCohort

(Optional) Numbers of maternal grandmother assignments by offspring cohort.

maternalGrandfathersByCohort

(Optional) Numbers of maternal grandmother assignments by offspring cohort.

paternalGrandmothersByCohort

(Optional) Numbers of paternal grandfather assignments by offspring cohort.

paternalGrandfathersByCohort

(Optional) Numbers of paternal grandfather assignments by offspring cohort.

cumulativePedigreeDepth

(Optional) Distributions of pedigree depth by cohort.

meanRelatednessAmongCohorts

(Optional) Mean relatedness among cohorts.

cohorts

(Optional) Returns cohort designations.

Graphical summaries of a number of these summary statistics are printed to the console when graphicalReports=='y'.

Examples

if (FALSE) { # \dontrun{

data(gryphons)
pedigree <- gryphons[, 1:3]

gryphons_ped_stats <- ped_stats(pedigree,
  cohorts = gryphons$cohort)

gryphons_ped_stats$totalMaternities
gryphons_ped_stats$paternitiesByCohort

summary(gryphons_ped_stats)

plot(gryphons_ped_stats)
} # }