A function to simulated phenotypic data

Simulates phenotypic data across arbitrary pedigrees. phen_sim simulate direct, maternal and paternal genetic and environmental effects for an arbitrary number of traits with arbitrary patterns of missing data.

Usage

phen_sim(
  pedigree,
  traits = 1,
  randomA = NULL,
  randomE = NULL,
  parentalA = NULL,
  parentalE = NULL,
  sampled = NULL,
  records = NULL,
  returnAllEffects = FALSE
)

Arguments

pedigree

A pedigree

traits

The number of traits for which data should be simulated.

randomA

An additive genetic covariance matrix, with dimensions a multiple of traits - see details

randomE

An additive environmental covariance matrix, with dimensions a multiple of traits - see details

parentalA

A vector indicating which effects in randomA (if any) to treat as parental effects

parentalE

A vector indicating which effects in randomE (if any) to treat as parental effects

sampled

A vector indicating which individuals are sampled

records

A single value, array of matrix specifying data record availability - see details

returnAllEffects

If TRUE then all individual breeding values and environmental effects are returned

Value

phenotypes

A dataframe containing phenotypes for all individuals specified to have records.

allEffects

(optional) A dataframe with all direct and indirect genetic and environmental effects.

Details

randomA and randomE are square matrices with dimension equal to the sum of the number direct and indirect effects. This must be a multiple of the number of traits, i.e. if an indirect effect is to be simulated for only one of multiple traits, those traits with no indirect effect should be included with (co)variances of zero.

parentalA and parentalE are optional vectors of characters indicating which trait positions in randomA and randomE are to be treated as indirect effects, and which effects to treat as maternal or paternal. Valid values are 'd', 'm', and 'p', for direct, maternal indirect and paternal indirect effects, respectively.

records can be specified either (1) as a single value to be applied to all individuals and traits, (2) as a vector the same length as the number of traits, representing trait-specific rates to be applied uniformly to all individuals, or (3) as data frames with rows for each individual and columns for each trait. In the third option, observed patterns of data availability can be simulated by supplying 0s and 1s for missing and available individual genotypes, respectively.

See also

Examples

## make up a pedigree
id <- c("a1", "a2", "a3", "a4", "a5", "a6", "a7", "a8", "a9")
dam <- c(NA, NA, NA, "a1", "a1", "a1", "a4", "a4", "a4")
sire <- c(NA, NA, NA, "a2", "a2", "a2", "a5", "a6", "a6")
pedigree <- as.data.frame(cbind(id, sire, dam))

traits <- 2
## no correlations
randomA <- diag(4)
randomE <- diag(4)
parentalA <- c("d", "d", "m", "m")
parentalE <- c("d", "d", "m", "m")

## generate phenoypic data based on this architecture
phen_sim(
  pedigree = pedigree, traits = 2, randomA = randomA, randomE = randomE,
  parentalA = parentalA, parentalE = parentalE
)
#> Simulating breeding values and environmental effects...done. 
#> Calculating phenotypes...done. 
#> $phenotypes
#>    id    trait_1    trait_2
#> a1 a1 -2.1543727  1.0496300
#> a2 a2  0.5121640 -0.1652518
#> a3 a3 -0.9382492  0.1022400
#> a4 a4  0.6400959 -0.2518556
#> a5 a5  0.3645492 -0.7251698
#> a6 a6  2.9096596  2.6040775
#> a7 a7 -0.7399189  1.3612447
#> a8 a8  0.4500141 -1.0180013
#> a9 a9  2.7681576  0.6924113
#> 

## let's do it again but see how the phenotypes were composed
phen_sim(
  pedigree = pedigree, traits = 2, randomA = randomA, randomE = randomE,
  parentalA = parentalA, parentalE = parentalE, returnAllEffects = TRUE
)
#> Simulating breeding values and environmental effects...done. 
#> Calculating phenotypes...done. 
#> $phenotypes
#>    id     trait_1     trait_2
#> a1 a1 -0.58712126  1.60206528
#> a2 a2 -4.37478225  1.21041313
#> a3 a3  0.56852801  0.73824510
#> a4 a4 -0.33852939  0.38431169
#> a5 a5  0.01098483 -0.08103798
#> a6 a6 -0.52486905  0.93101551
#> a7 a7  1.01660687 -1.61843254
#> a8 a8  0.05202055  1.18594418
#> a9 a9 -1.08385598  0.11114675
#> 
#> $allEffects
#>    id sire  dam       a_tr1    a_tr2   bv_m_tr1    bv_m_tr2       e_tr1
#> a1 a1 <NA> <NA> -0.30562067 2.211769 -1.0416684 -1.14652385  0.54214155
#> a2 a2 <NA> <NA> -1.67532730 1.525939  0.5541855  1.99311026 -1.45229977
#> a3 a3 <NA> <NA> -0.15412074 2.564408  1.0619991  1.14269488  0.04020439
#> a4 a4   a2   a1 -0.19579992 1.588132 -0.8258750  0.01395996  0.34042449
#> a5 a5   a2   a1  0.25671344 1.962894  0.0224838  1.22848111  0.23742535
#> a6 a6   a2   a1 -0.11276845 2.301668 -0.5473590  1.38528464  0.07105336
#> a7 a7   a5   a4 -0.01964701 1.583071 -2.1317596  0.66752659  1.15337579
#> a8 a8   a6   a4 -0.93104728 1.497175 -2.1458410  2.57270010  1.10018974
#> a9 a9   a6   a4 -0.96986001 1.704032 -0.1305747 -0.19876614  0.00312594
#>          e_tr2    e_m_tr1     e_m_tr2 P_bv_m_tr1  P_bv_m_tr2  P_e_m_tr1
#> a1  0.07510590  0.5585144  0.41540640  -1.113417 -0.20475662  0.2897745
#> a2  0.94120612 -0.3389359 -0.07557425  -1.396835  0.17658945  0.1496794
#> a3  0.12430107 -0.9984326  1.23339006   1.588785  0.15168817 -0.9063402
#> a4 -0.47270248  0.7087531 -1.52895871  -1.041668 -1.14652385  0.5585144
#> a5 -1.31281425  0.7470286 -1.56251843  -1.041668 -1.14652385  0.5585144
#> a6 -0.63953477 -0.8451957  0.67524470  -1.041668 -1.14652385  0.5585144
#> a7 -1.68650474 -0.9028149  1.31763370  -0.825875  0.01395996  0.7087531
#> a8  1.20376784 -1.4312708  1.38291086  -0.825875  0.01395996  0.7087531
#> a9 -0.07788682  0.4414282  0.12892290  -0.825875  0.01395996  0.7087531
#>     P_e_m_tr2    Phen_tr1    Phen_tr2
#> a1 -0.4800535 -0.58712126  1.60206528
#> a2 -1.4333211 -4.37478225  1.21041313
#> a3 -2.1021525  0.56852801  0.73824510
#> a4  0.4154064 -0.33852939  0.38431169
#> a5  0.4154064  0.01098483 -0.08103798
#> a6  0.4154064 -0.52486905  0.93101551
#> a7 -1.5289587  1.01660687 -1.61843254
#> a8 -1.5289587  0.05202055  1.18594418
#> a9 -1.5289587 -1.08385598  0.11114675
#>